cleanup
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@1812 98333e67-4a24-44d7-a75c-e53540dd3050

+++ trunk/include/xmlrpc-c/base_int.h
@@ -117,7 +117,7 @@ xmlrpc_typeName(xmlrpc_type const type);
 void
 xmlrpc_traceXml(const char * const label, 
                 const char * const xml,
-                unsigned int const xmlLength);
+                size_t       const xmlLength);
 
 void
 xmlrpc_destroyString(xmlrpc_value * const stringP);
